[infra] Extract validation visualization handlers
